Move AggResult from rulenode.rs to count.rs
AggResult is the result of a count/aggregation: it is produced by
count::aggregation_condition_select (the only place it is constructed in
production), consumed throughout count.rs (~50 refs), and its own
agg_record_time_info field is a Vec<count::AggRecordTimeInfo>. rulenode.rs
only returned it from RuleNode::judge_satisfy_aggcondition, which just
delegates to count::. Moving it into count.rs co-locates it with the
aggregation logic and the type it depends on, and removes the previous
rule<->count coupling (count.rs no longer imports AggResult from rule).

mod.rs re-exports it (`pub use count::AggResult;`) so the existing
`crate::detections::rule::AggResult` path is unchanged. Pure code
movement; detection output stays byte-identical and all tests pass.

This completes the split discussed in #1045 (per @fukusuket's suggestion
to house AggResult in count.rs).
