+++ CHANGELOG-Japanese.md
@@ -16,7 +16,7 @@
 
 **その他:**
 
-- `src/detections/rule/mod.rs`（モジュール宣言と、コアとなるルール型や大きなテストモジュールが混在した約1,200行）を、`RuleNode`・`DetectionNode`・`AggResult`・`CorrelationType`・`create_rule`/`get_detection_keys`・テストを収めた新しい `rulenode.rs` に分離した。`mod.rs` はモジュールの取りまとめと `pub use rulenode::*;` のみになり、既存のインポートパスはすべて変更なしで動作する。純粋なコード移動で、検知結果の出力はバイト単位で同一。 (#1890) (@YamatoSecurity)
+- `src/detections/rule/mod.rs`（モジュール宣言と、コアとなるルール型や大きなテストモジュールが混在した約1,200行）を、`RuleNode`・`DetectionNode`・`CorrelationType`・`create_rule`/`get_detection_keys`・テストを収めた新しい `rulenode.rs` に分離し、`AggResult` は（それを生成する集計ロジックと、依存先の `AggRecordTimeInfo` の隣である）`count.rs` に移動した。`mod.rs` はモジュールの取りまとめと再エクスポートのみになり、既存のインポートパスはすべて変更なしで動作する。純粋なコード移動で、検知結果の出力はバイト単位で同一。 (#1890) (@YamatoSecurity)
 - `src/main.rs` の約860行に及ぶ `App::analysis_files` 関数をリファクタリングした。この関数は対話式のルールセットウィザード全体、可読性の低い5つの位置引数を取るカウント用クロージャ、アドホックなチャンネルフィルタ、プログレスバーの構築をインラインで抱え込んでいた。4つの名前付きユニットを抽出した: `run_scan_wizard()`（対話式ウィザード全体。選択されたルールセットのラベルを返す）、`calculate_wizard_rule_count()`（先頭の `bool` と3つの `Vec<&str>` という位置引数の代わりに `WizardCountFilter` 構造体を取る自由関数に昇格したクロージャ。ユニットテスト付き）、`apply_channel_filters()`（`logon-summary`/`config-critical-systems`/`log-metrics` のアドホックなチャンネルフィルタ）、`build_progress_bar()`。`analysis_files` は、合計の表示・ウィザード・ルール読み込み・フィルタ・スキャンループ・出力、という流れで読めるようになった。挙動に変更はない — 影響を受けるすべてのサブコマンドで出力はバイト単位で同一であり、対話式ウィザードの経路（`--no-wizard` のスキャンでは通らない）は敵対的レビューによって忠実な行単位の移動であることを確認した。 (#1885) (@YamatoSecurity)
 - `src/detections/configs.rs` の約700行に及ぶ `StoredStatic::create_static_data` 関数をリファクタリングした。この関数は実際には決して `None` にならない `Option<Config>` を受け取り、それを33回 `.unwrap()` していた。シグネチャを `Config` を受け取るように変更し、アクションを一度だけ束縛する（`let action = config.action.as_ref()`）ようにして、繰り返しの unwrap と `None` でパニックする危険を取り除いた（すべての呼び出し側を更新）。また、コピペを解消するために3つのヘルパーを抽出した: 9個の設定ファイルで重複していた12行の設定ファイルフォールバックチェーンのための `resolve_config_file()`、繰り返されていた `LEVEL::iter().any(...)` のレベル検証のための `is_valid_level()`、include/exclude の computer・EID の `.iter().map(CompactString::from).collect()` という約120行のコピペのための `compact_string_set()`。`configs.rs` を正味 −250 行削減し、挙動に変更はない — 影響を受けるすべてのサブコマンド（`csv-timeline`、`json-timeline`、`logon-summary`、`computer-metrics`、`log-metrics`、`search`）で出力はバイト単位で同一（`eid-metrics` と `logon-summary` の成功テーブルも一致する。ただし未改変のバイナリにも存在するグループ内の同点時の並び順の非決定性を除く）。 (#1884) (@YamatoSecurity)
 - `search` コマンドのヒット処理ロジックの重複を解消した。`search_keyword` と `search_regex` は、マッチ判定の後に完全に同一な約65行の末尾処理（イベント情報の抽出、`Provider_EventID` によるフィールド略称テーブルの参照、AllFieldInfo の空白の正規化、そしてソート用にヒットを収集するか逐次書き出すかの分岐）を重複して持っており、ヒット処理への修正を2箇所に適用する必要があって、2つのコピーが乖離する恐れがあった。両方のループがマッチ判定の後に呼び出す単一の `emit_hit` メソッドに抽出した。出力は、キーワード検索・正規表現検索それぞれのソートあり/逐次出力の両方でバイト単位で同一。 (#1881) (@YamatoSecurity)
