+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 **Other:**
 
-- Split `src/detections/rule/mod.rs` (~1,200 lines that mixed the module declarations with the core rule types and a large test module) into a new `rulenode.rs` holding `RuleNode`, `DetectionNode`, `AggResult`, `CorrelationType`, `create_rule`/`get_detection_keys`, and the tests; `mod.rs` is now just module wiring plus `pub use rulenode::*;` so every existing import path is unchanged. Pure code movement with byte-identical detection output. (#1890) (@YamatoSecurity)
+- Split `src/detections/rule/mod.rs` (~1,200 lines that mixed the module declarations with the core rule types and a large test module) into a new `rulenode.rs` holding `RuleNode`, `DetectionNode`, `CorrelationType`, `create_rule`/`get_detection_keys`, and the tests, and moved `AggResult` to `count.rs` (next to the aggregation logic and the `AggRecordTimeInfo` it depends on); `mod.rs` is now just module wiring plus re-exports, so every existing import path is unchanged. Pure code movement with byte-identical detection output. (#1890) (@YamatoSecurity)
 - Refactored the ~860-line `App::analysis_files` function in `src/main.rs`, which inlined the entire interactive rule-set wizard, an unreadable five-positional-arg count closure, ad-hoc channel filters, and progress-bar setup. Extracted four named units: `run_scan_wizard()` (the whole interactive wizard, returning the selected rule-set label), `calculate_wizard_rule_count()` (the closure promoted to a free function taking a `WizardCountFilter` struct instead of a leading `bool` plus three `Vec<&str>`, now unit-tested), `apply_channel_filters()` (the `logon-summary`/`config-critical-systems`/`log-metrics` ad-hoc channel filters), and `build_progress_bar()`. `analysis_files` now reads as: print totals, wizard, load rules, filter, scan loop, emit. No behavior change — output is byte-identical across every affected subcommand, and the interactive wizard path (not covered by `--no-wizard` scans) was confirmed a faithful line-by-line move by an adversarial review. (#1885) (@YamatoSecurity)
 - Refactored the ~700-line `StoredStatic::create_static_data` function in `src/detections/configs.rs`. It took an `Option<Config>` that was never `None` and `.unwrap()`ed it 33 times; the signature now takes `Config` and binds the action once (`let action = config.action.as_ref()`), removing the repeated unwraps and the panic-on-`None` footgun (all call sites updated). Also extracted three helpers to collapse the copy-paste: `resolve_config_file()` for the 12-line config-file fallback chain that was duplicated for 9 setting files, `is_valid_level()` for the repeated `LEVEL::iter().any(...)` level validation, and `compact_string_set()` for the ~120 lines of include/exclude computer and EID `.iter().map(CompactString::from).collect()` copy-paste. Net −250 lines in `configs.rs` with no behavior change — output is byte-identical across every affected subcommand (`csv-timeline`, `json-timeline`, `logon-summary`, `computer-metrics`, `log-metrics`, `search`; `eid-metrics` and the `logon-summary` successful table match too, modulo an intra-group tie-order nondeterminism present in the unmodified binary). (#1884) (@YamatoSecurity)
 - De-duplicated the search command's hit-processing logic. `search_keyword` and `search_regex` carried a verbatim-identical ~65-line tail (extract the event info, look up the `Provider_EventID` field-abbreviation table, normalize the AllFieldInfo whitespace, and either collect the hit for sorting or write it out on the fly), so any fix to the hit path had to be applied twice and the two copies could drift. It is now a single `emit_hit` method that both loops call after their match predicate. Output is byte-identical for both search modes, sorted and on-the-fly. (#1881) (@YamatoSecurity)
